backend/login/requirements.txt
flask
mysql-connector-python
flask_cors
uuid